Support broader markdown and whitespace formatting
By experimentation, GFM seems to mostly support the following:
 - Any combination of zero or more space (\u0020) and tab (\u0009)
   characters to indent from the beginning of the line to the list
   bullet character
 - Any of '-', '*', or '+' as a bullet character
 - Any combination of one or more space (\u0020) and tab (\u0009)
   characters to separate the bullet from the opening bracket '['
 - A single whitespace character (regex \s character class) within
   the brackets '[' and ']' to indicate unchecked, or either 'x' or
   'X' to indicate checked.

However there are some exceptions to this rule, e.g. this sequence
is NOT interpreted as a task list item: "-\u0009[\u0009] Item".

For the sake of simplicity, we shall not attempt to mimic this
behavior precisely, but instead generously allow arbitrary sequences
of spaces and tabs.